feat: ensure notes and page break are stored as embed blots and simplify keyboard navigation
There is a lot of stuff in there.

Main changes:

RichTextEditor is now initialized with its initial value props. After
that, it is not controlled and only sends its new value to its parent
component in the hierarchy.

NoteForm is updated so that the first time it is rendered, the
RichTextEditor value props is already defined. Before this change, the
editor was populated on second render, after the call to the mounted
lifecycle hook.

Components that take input data as propse now expect a single data
structure as prop. For example, NoteForm `inputData` prop could be
either {note: {id: number}} or {formats: {note: {id: number}}, leading
to big conditionals.

A lot of imperative logic is transformed into more functional logic by
using computed component state. This is the case for example for the
way the "activeness" of the buttons is computed. We remove the
`updateButtons` method from the RichTextEditor and instead compute the
buttons availability from the component's current state on render.

Notes and Page breaks are transformed into embeds to avoid hard to
maintain logic to ensure that their inner text are not modified.
This allow us to write specific keyboard and mouse handlers for these
elements the embeds are selected on click ot when navigating over them
with keyboard.

A new `remove format` button is added in the RichTextEditor which allows
to ensure that the text going to be typed will not be formatted. A new
icon should be chosen for this button as it is currently the same as
"stroked text"/"del" button.